backend/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
PyJWT==2.8.0
Pillow==10.3.0